+++ lib/tcr/actions/tcr_action_disable.sh
@@ -8,11 +8,17 @@ source "$TCR_HOME/lib/tcr/actions/tcr_action_watch.sh"
 TCR_ACTION_DISABLE='stop'
 
 tcr_action_disable() {
-    lock_file_remove
-
     if tcr_action_watch_is_running; then
         tcr_action_watch_stop
     fi
 
-    print_status 'OFF'
+    session_name="$(tcr_session_name)"
+
+    lock_file_remove
+
+    if is_set "$session_name"; then
+        print_status "session '$session_name' stopped"
+    else
+        print_status 'session stopped'
+    fi
 }